Redo TimePicker components and example
