+++ docs/pages/api/timepicker/InlineTimePicker.example.jsx
@@ -1,5 +1,5 @@
 import React, { Fragment, useState } from 'react';
-import { InlineTimePicker } from 'material-ui-pickers';
+import { TimePicker, KeyboardTimePicker } from 'material-ui-pickers';
 
 function InlineTimePickerDemo() {
   const [selectedDate, handleDateChange] = useState('2018-01-01T00:00:00.000Z');
@@ -7,12 +7,17 @@ function InlineTimePickerDemo() {
   return (
     <Fragment>
       <div className="picker">
-        <InlineTimePicker label="Inline mode" value={selectedDate} onChange={handleDateChange} />
+        <TimePicker
+          variant="inline"
+          label="Inline mode"
+          value={selectedDate}
+          onChange={handleDateChange}
+        />
       </div>
 
       <div className="picker">
-        <InlineTimePicker
-          keyboard
+        <KeyboardTimePicker
+          variant="inline"
           label="With keyboard"
           value={selectedDate}
           onChange={handleDateChange}
