+++ lib/src/DatePicker/KeyboardDatePicker.tsx
@@ -8,7 +8,6 @@ import {
 import { useUtils } from '../_shared/hooks/useUtils';
 import KeyboardDateInput, { KeyboardDateInputProps } from '../_shared/KeyboardDateInput';
 import { ExtendWrapper2, Wrapper } from '../wrappers/Wrapper';
-import { DatePicker } from './DatePicker';
 import DatePickerRoot, { BaseDatePickerProps } from './DatePickerRoot';
 
 export type KeyboardDatePickerProps = BaseDatePickerProps &
@@ -82,7 +81,7 @@ export function KeyboardDatePicker(props: KeyboardDatePickerProps) {
   );
 }
 
-KeyboardDatePicker.defaultProps = DatePicker.defaultProps;
+KeyboardDatePicker.defaultProps = DatePickerRoot.defaultProps;
 
 export default React.forwardRef((props: KeyboardDatePickerProps, ref) => (
   <KeyboardDatePicker {...props} forwardedRef={ref} />
