+++ lib/src/DatePicker/index.ts
@@ -1,5 +1,10 @@
 import { DatePickerProps } from './DatePicker';
+import { KeyboardDatePickerProps } from './KeyboardDatePicker';
 
 export type DatePickerProps = DatePickerProps;
+
+export type KeyboardDatePickerProps = KeyboardDatePickerProps;
+
+export { KeyboardDatePicker } from './KeyboardDatePicker';
 
 export { default } from './DatePicker';
