+++ lib/src/_shared/BasePicker.tsx
@@ -34,8 +34,6 @@ export interface BasePickerProps {
   format?: string;
   /** Dynamic formatter of text field label */
   labelFunc?: (date: MaterialUiPickersDate, invalidLabel: string) => string;
-  /** Do not open picker on enter keypress */
-  disableOpenOnEnter?: boolean;
   /** Date that will be initially highlighted */
   initialFocusedDate?: DateType;
   /** Message displaying in text field, if date is invalid (doesn't work in keyboard mode) */
